+++ internal/component/metrics/namespaced_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/benthosdev/benthos/v4/internal/bundle"
 	"github.com/benthosdev/benthos/v4/internal/component/metrics"
 	"github.com/benthosdev/benthos/v4/internal/log"
+	"github.com/benthosdev/benthos/v4/internal/manager/mock"
 
 	_ "github.com/benthosdev/benthos/v4/public/components/all"
 )
@@ -23,7 +24,7 @@ func getTestProm(t *testing.T) (metrics.Type, http.HandlerFunc) {
 	conf.Type = "prometheus"
 	conf.Prometheus.UseHistogramTiming = true
 
-	ns, err := bundle.AllMetrics.Init(conf, log.Noop())
+	ns, err := bundle.AllMetrics.Init(conf, mock.NewManager())
 	require.NoError(t, err)
 
 	prom := ns.Child()