+++ internal/component/tracer/config.go
@@ -22,6 +22,7 @@ type Config struct {
 	Jaeger     JaegerConfig     `json:"jaeger" yaml:"jaeger"`
 	CloudTrace CloudTraceConfig `json:"gcp_cloudtrace" yaml:"gcp_cloudtrace"`
 	None       struct{}         `json:"none" yaml:"none"`
+	Plugin     interface{}      `json:"plugin,omitempty" yaml:"plugin,omitempty"`
 }
 
 // NewConfig returns a configuration struct fully populated with default values.
@@ -31,6 +32,7 @@ func NewConfig() Config {
 		Jaeger:     NewJaegerConfig(),
 		CloudTrace: NewCloudTraceConfig(),
 		None:       struct{}{},
+		Plugin:     nil,
 	}
 }
 
@@ -47,8 +49,19 @@ func (conf *Config) UnmarshalYAML(value *yaml.Node) error {
 		return fmt.Errorf("line %v: %v", value.Line, err)
 	}
 
-	if aliased.Type, _, err = docs.GetInferenceCandidateFromYAML(docs.DeprecatedProvider, docs.TypeTracer, value); err != nil {
+	var spec docs.ComponentSpec
+	if aliased.Type, spec, err = docs.GetInferenceCandidateFromYAML(docs.DeprecatedProvider, docs.TypeTracer, value); err != nil {
 		return fmt.Errorf("line %v: %w", value.Line, err)
+	}
+
+	if spec.Plugin {
+		pluginNode, err := docs.GetPluginConfigYAML(aliased.Type, value)
+		if err != nil {
+			return fmt.Errorf("line %v: %v", value.Line, err)
+		}
+		aliased.Plugin = &pluginNode
+	} else {
+		aliased.Plugin = nil
 	}
 
 	*conf = Config(aliased)