+++ internal/impl/aws/metrics_cloudwatch.go
@@ -20,8 +20,8 @@ import (
 )
 
 func init() {
-	_ = bundle.AllMetrics.Add(func(c metrics.Config, l log.Modular) (metrics.Type, error) {
-		return newCloudWatch(c.AWSCloudWatch, l)
+	_ = bundle.AllMetrics.Add(func(c metrics.Config, nm bundle.NewManagement) (metrics.Type, error) {
+		return newCloudWatch(c.AWSCloudWatch, nm.Logger())
 	}, docs.ComponentSpec{
 		Name:    "aws_cloudwatch",
 		Type:    docs.TypeMetrics,