+++ internal/impl/pure/metrics_none.go
@@ -4,11 +4,10 @@ import (
 	"github.com/benthosdev/benthos/v4/internal/bundle"
 	"github.com/benthosdev/benthos/v4/internal/component/metrics"
 	"github.com/benthosdev/benthos/v4/internal/docs"
-	"github.com/benthosdev/benthos/v4/internal/log"
 )
 
 func init() {
-	_ = bundle.AllMetrics.Add(func(metrics.Config, log.Modular) (metrics.Type, error) {
+	_ = bundle.AllMetrics.Add(func(metrics.Config, bundle.NewManagement) (metrics.Type, error) {
 		return metrics.Noop(), nil
 	}, docs.ComponentSpec{
 		Name:    "none",