+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - The `redis` processor has been reworked to be more generally useful, the old `operator` and `key` fields are now deprecated in favour of new `command` and `args_mapping` fields.
 - Go API: Added component bundle `./public/components/aws` for all AWS components, including a `RunLambda` function.
 - New `cached` processor.
+- Go API: New APIs for registering both metrics exporters and open telemetry tracer plugins.
 
 ### Fixed